Move the html parser config argument to the specific job which uses it
